Personalized_Shopping _Recommendation/backend/requirements.txt
fastapi
uvicorn
pandas
torch
transformers
python-dotenv
huggingface_hub
langchain
langchain-community
langchain-huggingface